feat: split bot logic
